+++ src/test/kotlin/octris/forgejo/MyPluginTest.kt
@@ -1,4 +1,4 @@
-package com.github.nicba1010.intellijplatformforgejoplugin
+package octris.forgejo
 
 import com.intellij.ide.highlighter.XmlFileType
 import com.intellij.openapi.components.service
@@ -6,7 +6,7 @@ import com.intellij.psi.xml.XmlFile
 import com.intellij.testFramework.TestDataPath
 import com.intellij.testFramework.fixtures.BasePlatformTestCase
 import com.intellij.util.PsiErrorElementUtil
-import com.github.nicba1010.intellijplatformforgejoplugin.services.MyProjectService
+import octris.forgejo.services.MyProjectService
 
 @TestDataPath("\$CONTENT_ROOT/src/test/testData")
 class MyPluginTest : BasePlatformTestCase() {